split module and bug fix in deleting behavior
